Added the Router to Bookshelf
